Move to the collection package
