+++ service/matching/matching_engine.go
@@ -58,7 +58,7 @@ import (
 	"go.temporal.io/server/common/clock"
 	hlc "go.temporal.io/server/common/clock/hybrid_logical_clock"
 	"go.temporal.io/server/common/cluster"
-	"go.temporal.io/server/common/concurrent"
+	"go.temporal.io/server/common/collection"
 	"go.temporal.io/server/common/dynamicconfig"
 	"go.temporal.io/server/common/log"
 	"go.temporal.io/server/common/log/tag"
@@ -135,17 +135,17 @@ type (
 		// that QueryWorkflow() will block on. The channel is unblocked either by worker sending response through
 		// RespondQueryTaskCompleted() or through an internal service error causing temporal to be unable to dispatch
 		// query task to workflow worker.
-		queryResults concurrent.Map[string, chan *queryResult]
+		queryResults collection.SyncMap[string, chan *queryResult]
 		// nexusResults maps nexus TaskID (which is a UUID generated in the DispatchNexusTask() call) to
 		// a channel that DispatchNexusTask() blocks on. The channel is unblocked either by worker responding
 		// via RespondNexusTaskCompleted() or RespondNexusTaskFailed(), or through an internal service error.
-		nexusResults concurrent.Map[string, chan *nexusResult]
+		nexusResults collection.SyncMap[string, chan *nexusResult]
 		// outstandingPollers is needed to keep track of all outstanding pollers for a particular
 		// taskqueue.  PollerID generated by frontend is used as the key and CancelFunc is the
 		// value.  This is used to cancel the context to unblock any outstanding poller when
 		// the frontend detects client connection is closed to prevent tasks being dispatched
 		// to zombie pollers.
-		outstandingPollers concurrent.Map[string, context.CancelFunc]
+		outstandingPollers collection.SyncMap[string, context.CancelFunc]
 		// Only set if global namespaces are enabled on the cluster.
 		namespaceReplicationQueue persistence.NamespaceReplicationQueue
 		// Disables concurrent task queue user data updates and replication requests (due to a cassandra limitation)
@@ -207,9 +207,9 @@ func NewEngine(
 		taskQueues:                make(map[taskQueueID]taskQueueManager),
 		taskQueueCount:            make(map[taskQueueCounterKey]int),
 		config:                    config,
-		queryResults:              concurrent.NewMap[string, chan *queryResult](),
-		nexusResults:              concurrent.NewMap[string, chan *nexusResult](),
-		outstandingPollers:        concurrent.NewMap[string, context.CancelFunc](),
+		queryResults:              collection.NewSyncMap[string, chan *queryResult](),
+		nexusResults:              collection.NewSyncMap[string, chan *nexusResult](),
+		outstandingPollers:        collection.NewSyncMap[string, context.CancelFunc](),
 		namespaceReplicationQueue: namespaceReplicationQueue,
 		namespaceUpdateLockMap:    make(map[string]*namespaceUpdateLocks),
 	}
