+++ service/matching/matching_engine_test.go
@@ -67,7 +67,7 @@ import (
 	"go.temporal.io/server/common/clock"
 	hlc "go.temporal.io/server/common/clock/hybrid_logical_clock"
 	"go.temporal.io/server/common/cluster"
-	"go.temporal.io/server/common/concurrent"
+	"go.temporal.io/server/common/collection"
 	"go.temporal.io/server/common/dynamicconfig"
 	"go.temporal.io/server/common/log"
 	"go.temporal.io/server/common/log/tag"
@@ -173,7 +173,7 @@ func newMatchingEngine(
 		historyClient:       mockHistoryClient,
 		taskQueues:          make(map[taskQueueID]taskQueueManager),
 		taskQueueCount:      make(map[taskQueueCounterKey]int),
-		queryResults:        concurrent.NewMap[string, chan *queryResult](),
+		queryResults:        collection.NewSyncMap[string, chan *queryResult](),
 		logger:              logger,
 		throttledLogger:     log.ThrottledLogger(logger),
 		metricsHandler:      metrics.NoopMetricsHandler,
